+++ internal/consul/certificates/certuri/uri_service.go
@@ -1,4 +1,4 @@
-package uri
+package certuri
 
 import (
 	"fmt"
