+++ internal/consul/certificates/certuri/uri_service_test.go
@@ -1,4 +1,4 @@
-package uri
+package certuri
 
 import (
 	"testing"
