+++ internal/consul/certificates/certuri/uri_signing_test.go
@@ -1,4 +1,4 @@
-package uri
+package certuri
 
 import (
 	"net/url"
